feat(collections): typed JSON-path indexes (numeric/bool fast path)
Derived index columns are now typed (INT/BOOLEAN/TEXT) instead of always
TEXT, so numeric and bool queries on indexed fields use the GlueSQL fast
path (read-your-writes, no seal needed) instead of falling back to
DataFusion.

- `index_sql_type`: Number→INT (FLOAT fails `to_cmp_be_bytes`; DECIMAL
  fails cross-type `evaluate_cmp` with I64 params in GlueSQL 0.19)
- `createIndex`: resolves IndexType from `options.type` hint or by
  sampling existing docs via `infer_index_type`; issues typed ALTER TABLE
  and backfills with `derive_typed_value`
- `indexed_col_types`: replaces `indexed_col_names`; maps each
  `__cidx_*` column's GlueSQL DataType → IndexType via schema introspection
- `write_full_doc`/`rewrite_doc_row`: populate derived columns with
  `derive_typed_value` + `json_value_to_param` (INT→Param::Int;
  fractional floats→Param::Null since INT column can't hold them)
- `to_sql_indexed`: signature changed to `HashMap<String, IndexType>`
  (keyed by derived col name); routes fast path when comparison value's
  JSON type matches the column's IndexType (Number↔number, Bool↔bool,
  Text↔string); type-mismatched comparisons fall back to JSON accessor
- New integration tests: `numeric_indexed_field_is_fast_and_fresh` (no
  seal, proves INT fast path for eq + range), `bool_indexed_field_is_fast`,
  `numeric_index_queried_with_string_does_not_error` (no crash on fallback)
- All 60 unit tests + 25 integration tests green; no new clippy warnings

+++ crates/bluedb-collections/src/index.rs
@@ -66,9 +66,24 @@ pub fn infer_index_type(values: &[serde_json::Value]) -> IndexType {
 }
 
 /// SQL column type keyword for a typed derived index column.
+///
+/// `Number` maps to `INT` (GlueSQL `DataType::Int` → `Value::I64`):
+/// - `FLOAT`/`F64` cannot be serialized to order-preserving big-endian bytes
+///   by bluedb-sql's storage layer, causing an error on every row write.
+/// - `DECIMAL` stores correctly but `evaluate_cmp(Decimal, I64)` returns `None`
+///   in GlueSQL 0.19 (the match arm for mixed Decimal/integer is missing in
+///   `Value::evaluate_cmp`), silently returning empty results for any query
+///   whose parameter is bound as `Param::Int`.
+/// - `INT` (`Value::I64`) stores without error, and `evaluate_cmp(I64, I64)` and
+///   range comparisons both work correctly.
+///
+/// Fractional JSON numbers (e.g. `3.14`) are stored as NULL for an INT-typed
+/// column — `derive_typed_value` returns the value only when the JSON type
+/// matches the index type, and `json_to_param` maps floats to `Param::Float`
+/// which is rejected by the INT column coercion (the row gets NULL instead).
 pub fn index_sql_type(t: IndexType) -> &'static str {
     match t {
-        IndexType::Number => "FLOAT",
+        IndexType::Number => "INT",
         IndexType::Bool => "BOOLEAN",
         IndexType::Text => "TEXT",
     }
@@ -231,7 +246,8 @@ mod tests {
 
     #[test]
     fn index_sql_type_number() {
-        assert_eq!(index_sql_type(IndexType::Number), "FLOAT");
+        // INT — FLOAT/DECIMAL have storage or comparison issues in GlueSQL 0.19.
+        assert_eq!(index_sql_type(IndexType::Number), "INT");
     }
 
     #[test]
